requirements.txt
flask
pyyaml
ansible
PyPowerFlex
